Moved import audio dialog
